@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import warnings
-from typing import List, Literal
+from typing import Literal
 
 import torch
 
@@ -141,7 +141,7 @@ def disjoint_merge(task_tensors: torch.Tensor, majority_sign_mask: torch.Tensor)
     return mixed_task_tensors / torch.clamp(num_params_preserved, min=1.0)
 
 
-def task_arithmetic(task_tensors: List[torch.Tensor], weights: torch.Tensor) -> torch.Tensor:
+def task_arithmetic(task_tensors: list[torch.Tensor], weights: torch.Tensor) -> torch.Tensor:
     """
     Merge the task tensors using `task arithmetic`.
 
@@ -160,7 +160,7 @@ def task_arithmetic(task_tensors: List[torch.Tensor], weights: torch.Tensor) -> 
     return mixed_task_tensors
 
 
-def magnitude_prune(task_tensors: List[torch.Tensor], weights: torch.Tensor, density: float) -> torch.Tensor:
+def magnitude_prune(task_tensors: list[torch.Tensor], weights: torch.Tensor, density: float) -> torch.Tensor:
     """
     Merge the task tensors using `task arithmetic`.
 
@@ -183,7 +183,7 @@ def magnitude_prune(task_tensors: List[torch.Tensor], weights: torch.Tensor, den
 
 
 def ties(
-    task_tensors: List[torch.Tensor],
+    task_tensors: list[torch.Tensor],
     weights: torch.Tensor,
     density: float,
     majority_sign_method: Literal["total", "frequency"] = "total",
@@ -214,7 +214,7 @@ def ties(
     return mixed_task_tensors
 
 
-def dare_linear(task_tensors: List[torch.Tensor], weights: torch.Tensor, density: float) -> torch.Tensor:
+def dare_linear(task_tensors: list[torch.Tensor], weights: torch.Tensor, density: float) -> torch.Tensor:
     """
     Merge the task tensors using `dare linear`.
 
@@ -237,7 +237,7 @@ def dare_linear(task_tensors: List[torch.Tensor], weights: torch.Tensor, density
 
 
 def dare_ties(
-    task_tensors: List[torch.Tensor],
+    task_tensors: list[torch.Tensor],
     weights: torch.Tensor,
     density: float,
     majority_sign_method: Literal["total", "frequency"] = "total",
